refactor: 불필요한 찾기 관련 파일 삭제
